+++ admin/notice.php
@@ -21,14 +21,14 @@ function ufc_rating_admin_notice() {
         return;
     }
 
-    $dismiss = wp_nonce_url( add_query_arg( 'ufc_rating_notice_action', 'dismiss_rating_true' ), 'dismiss_rating_true' ); 
-    $no_thanks = wp_nonce_url( add_query_arg( 'ufc_rating_notice_action', 'no_thanks_rating_true' ), 'no_thanks_rating_true' ); ?>
+    $dismiss = wp_nonce_url( add_query_arg( 'ufc_rating_notice_action', 'ufc_dismiss_rating_true' ), 'ufc_dismiss_rating_true' ); 
+    $no_thanks = wp_nonce_url( add_query_arg( 'ufc_rating_notice_action', 'ufc_no_thanks_rating_true' ), 'ufc_no_thanks_rating_true' ); ?>
     
     <div class="notice notice-success">
-        <p><?php _e( 'Hey, I noticed you\'ve been using Ultimate Facebook Comments for more than 1 week – that’s awesome! Could you please do me a BIG favor and give it a <strong>5-star</strong> rating on WordPress? Just to help me spread the word and boost my motivation.', 'ultimate-facebook-comments' ); ?><p>
-        <a href="https://wordpress.org/support/plugin/ultimate-facebook-comments/reviews/?filter=5#new-post" target="_blank" class="button button-secondary"><?php _e( 'Ok, you deserve it', 'ultimate-facebook-comments' ); ?></a>&nbsp;
+        <p><?php _e( 'Hey, I noticed you\'ve been using Ultimate Facebook Comments for more than 1 week – that’s awesome! Could you please do me a BIG favor and give it a <strong>5-star</strong> rating on WordPress? Just to help me spread the word and boost my motivation.', 'ultimate-facebook-comments' ); ?></p>
+        <p><a href="https://wordpress.org/support/plugin/ultimate-facebook-comments/reviews/?filter=5#new-post" target="_blank" class="button button-secondary"><?php _e( 'Ok, you deserve it', 'ultimate-facebook-comments' ); ?></a>&nbsp;
         <a href="<?php echo $dismiss; ?>" class="already-did"><strong><?php _e( 'I already did', 'ultimate-facebook-comments' ); ?></strong></a>&nbsp;<strong>|</strong>
-        <a href="<?php echo $no_thanks; ?>" class="later"><strong><?php _e( 'Nope&#44; maybe later', 'ultimate-facebook-comments' ); ?></strong></a>
+        <a href="<?php echo $no_thanks; ?>" class="later"><strong><?php _e( 'Nope&#44; maybe later', 'ultimate-facebook-comments' ); ?></strong></a></p>
     </div>
 <?php
 }
@@ -47,13 +47,13 @@ function ufc_dismiss_rating_admin_notice() {
         return;
     }
 
-    if ( 'dismiss_rating_true' === $_GET['ufc_rating_notice_action'] ) {
-        check_admin_referer( 'dismiss_rating_true' );
+    if ( 'ufc_dismiss_rating_true' === $_GET['ufc_rating_notice_action'] ) {
+        check_admin_referer( 'ufc_dismiss_rating_true' );
         update_option( 'ufc_plugin_dismiss_rating_notice', '1' );
     }
 
-    if ( 'no_thanks_rating_true' === $_GET['ufc_rating_notice_action'] ) {
-        check_admin_referer( 'no_thanks_rating_true' );
+    if ( 'ufc_no_thanks_rating_true' === $_GET['ufc_rating_notice_action'] ) {
+        check_admin_referer( 'ufc_no_thanks_rating_true' );
         update_option( 'ufc_plugin_no_thanks_rating_notice', '1' );
         update_option( 'ufc_plugin_dismiss_rating_notice', '1' );
         update_option( 'ufc_plugin_dismissed_time', time() );