+++ ultimate-facebook-comments.php
@@ -2,8 +2,8 @@
 /**
  * Plugin Name: Ultimate Facebook Comments
  * Plugin URI: https://iamsayan.github.io/ultimate-facebook-comments/
- * Description: Ultimate Facebook Comments plugin will help you to display Facebook Comments box on your website easily. You can use Facebook Comments on your posts or pages.
- * Version: 1.4.1
+ * Description: 🔥 Ultimate Facebook Comments plugin will help you to display Facebook Comments box on your website easily. You can use Facebook Comments on your posts or pages.
+ * Version: 1.4.2
  * Author: Sayan Datta
  * Author URI: https://sayandatta.com
  * License: GPLv3
@@ -35,7 +35,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'UFC_PLUGIN_VERSION', '1.4.0' );
+define( 'UFC_PLUGIN_VERSION', '1.4.2' );
 
 // debug scripts
 //define ( 'UFC_PLUGIN_ENABLE_DEBUG', 'true' );
@@ -55,7 +55,6 @@ register_activation_hook( __FILE__, 'ufc_plugin_run_on_activation' );
 register_deactivation_hook( __FILE__, 'ufc_plugin_run_on_deactivation' );
 
 function ufc_plugin_run_on_activation() {
-
     if ( ! current_user_can( 'activate_plugins' ) ) {
         return;
     }
@@ -63,7 +62,6 @@ function ufc_plugin_run_on_activation() {
 }
 
 function ufc_plugin_run_on_deactivation() {
-
     if ( ! current_user_can( 'activate_plugins' ) ) {
         return;
     }
@@ -75,7 +73,6 @@ function ufc_plugin_run_on_deactivation() {
 
 //add admin styles and scripts
 function ufc_custom_admin_styles_scripts( $hook ) {
-
     $ver = UFC_PLUGIN_VERSION;
     if( defined( 'UFC_PLUGIN_ENABLE_DEBUG' ) ) {
         $ver = time();
@@ -83,11 +80,10 @@ function ufc_custom_admin_styles_scripts( $hook ) {
 
     $current_screen = get_current_screen();
     if ( strpos( $current_screen->base, 'ultimate-facebook-comments') !== false ) {
-        wp_enqueue_style( 'ufc-admin-style', plugins_url( 'admin/assets/css/admin.min.css', __FILE__ ), array(), $ver );
-        wp_enqueue_style( 'ufc-style', plugins_url( 'admin/assets/css/style.min.css', __FILE__ ), array(), $ver );      
+        wp_enqueue_style( 'ufc-admin', plugins_url( 'admin/assets/css/admin.min.css', __FILE__ ), array(), $ver );
         wp_enqueue_style( 'ufc-selectize-theme-css', plugins_url( 'admin/assets/css/selectize.css', __FILE__ ), array(), '0.12.6' );       
         
-        wp_enqueue_script( 'ufc-script', plugins_url( 'admin/assets/js/admin.min.js', __FILE__ ), array( 'jquery', 'wp-color-picker' ), $ver, true );
+        wp_enqueue_script( 'ufc-admin-js', plugins_url( 'admin/assets/js/admin.min.js', __FILE__ ), array( 'jquery', 'wp-color-picker' ), $ver, true );
         wp_enqueue_script( 'ufc-selectize-js', plugins_url( 'admin/assets/js/selectize.min.js', __FILE__ ), array(), '0.12.6' );
        
         wp_enqueue_style( 'wp-color-picker' );
@@ -149,40 +145,12 @@ function ufc_ajax_save_admin_scripts() {
 
 add_action( 'admin_init', 'ufc_ajax_save_admin_scripts' );
 
-function ufc_plugin_all_notices() { 
-    $options = get_option('ufc_plugin_global_options');
-
-    /* Check transient, if available display notice */
-    if( get_transient( 'ufc-admin-notice-on-activation' ) && empty( $options['ufc_facebook_comments_app_id'] ) ) { ?>
-        <div class="notice notice-success is-dismissible">
-            <p><strong><?php printf( __( 'Thanks for installing %1$s v%2$s plugin. Click <a href="%3$s">here</a> to configure plugin settings.', 'ultimate-facebook-comments' ), 'Ultimate Facebook Comments', UFC_PLUGIN_VERSION, admin_url( 'admin.php?page=ultimate-facebook-comments' ) ); ?></strong></p>
-        </div> <?php
-        /* Delete transient, only display this notice once. */
-        delete_transient( 'ufc-admin-notice-on-activation' );
-    }
-
-    if( empty( $options['ufc_facebook_comments_app_id'] ) && empty( $options['ufc_facebook_comments_app_secret'] ) ) { ?>
-        <div id="appid-notice" class="notice notice-error">
-            <p><?php _e( 'Please create and enter your Facebook App ID for Ultimate Facebook Comments Plugin to work properly. If you have already created an app, then put <strong>APP ID</strong> and <strong>APP Secret</strong> in plugin settings.', 'ultimate-facebook-comments' ); ?></p>
-            <p><a href="https://developers.facebook.com/apps" target="_blank" class="button button-secondary"><?php _e( 'Create New Facebook App ID', 'ultimate-facebook-comments' ); ?></a></p>
-        </div><?php
-    } elseif( !empty( $options['ufc_facebook_comments_app_id'] ) && empty( $options['ufc_facebook_comments_app_secret'] ) ) { ?>
-        <div id="appid-notice" class="notice notice-error">
-            <p><?php _e( 'Please enter your Facebook App Secret for Ultimate Facebook Comments Plugin to work properly. Go to Facebook for Developers and copy App Secret and paste it to plugin settings.', 'ultimate-facebook-comments' ); ?></p>
-            <p><a href="https://developers.facebook.com/apps/<?php echo $options['ufc_facebook_comments_app_id']; ?>/settings/basic/" target="_blank" class="button button-secondary"><?php _e( 'View App Secret', 'ultimate-facebook-comments' ); ?></a></p>
-        </div><?php
-    }
-}
-
-add_action( 'admin_notices', 'ufc_plugin_all_notices' ); 
-
-$options = get_option('ufc_plugin_global_options');
-
 require_once plugin_dir_path( __FILE__ ) . 'admin/loader.php';
 require_once plugin_dir_path( __FILE__ ) . 'admin/notice.php';
 require_once plugin_dir_path( __FILE__ ) . 'admin/donate.php';
+require_once plugin_dir_path( __FILE__ ) . 'admin/install.php';
 require_once plugin_dir_path( __FILE__ ) . 'admin/ajax.php';
-require_once plugin_dir_path( __FILE__ ) . 'admin/post-meta.php';
+require_once plugin_dir_path( __FILE__ ) . 'admin/update.php';
 require_once plugin_dir_path( __FILE__ ) . 'admin/disable-native.php';
 require_once plugin_dir_path( __FILE__ ) . 'admin/admin-bar.php';
 require_once plugin_dir_path( __FILE__ ) . 'admin/dashboard-edit-screen.php';